requirements.txt
arrow==1.4.0
asgiref==3.9.1
binaryornot==0.4.4
build==1.3.0
certifi==2026.1.4
chardet==5.2.0
charset-normalizer==3.4.4
click==8.3.1
cookiecutter==2.6.0
cssselect2==0.8.0
Django==5.2.6
django-appconf==1.2.0
django-browser-reload==1.21.0
django-debug-toolbar==6.0.0
django-js-asset==3.1.2
django-tailwind-4==0.1.4
django-tinymce==5.0.0
django_compressor==4.6.0
easy-thumbnails==2.10.1
freetype-py==2.5.1
honcho==2.0.0
idna==3.11
Jinja2==3.1.6
lxml==6.0.2
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
packaging==25.0
pillow==12.1.0
psycopg2==2.9.10
pycairo==1.29.0
Pygments==2.19.2
pyproject_hooks==1.2.0
python-dateutil==2.9.0.post0
python-slugify==8.0.4
PyYAML==6.0.3
rcssmin==1.2.2
reportlab==4.4.7
requests==2.32.5
rich==14.2.0
rjsmin==1.2.5
rlPyCairo==0.4.0
six==1.17.0
sqlparse==0.5.3
svglib==1.6.0
text-unidecode==1.3
tinycss2==1.5.1
tzdata==2025.3
urllib3==2.6.2
webencodings==0.5.1
